more dev

+++ src/test/java/com/colak/netty/udprpc/SolarisKey.java
@@ -1,7 +1,7 @@
 package com.colak.netty.udprpc;
 
 public record SolarisKey(
-        short protocolNo,
-        short subType,
-        short messageNo) {
+        int protocolNo,
+        int subType,
+        int messageNo) {
 }
